refactor: add schema files and rename views to blueprints
