+++ backend/bamboo/blueprints/admin.py
