+++ backend/bamboo/blueprints/talk.py
