--- backend/bamboo/schemas/__init__.py
@@ -1,1 +1,0 @@
-from apiflask import Schema, fields, validators
