+++ backend/bamboo/schemas/blog.py
@@ -1,0 +1,1 @@
+from apiflask import Schema, fields, validators
